+++ CHANGELOG.md
@@ -5,6 +5,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+
+## Upcoming
+
+### Added
+
+- Traktor config option `backup_before_write`, (true by default, creates a backup of the nml file before each write).
+
 ## [0.4.0] - 2026-03-07
 
 ### Added